requirements.txt
Django~=4.2.13
pillow==10.3.0
django-admin-sortable2==2.2.1
geojson==3.1.0
django-tinymce==4.0.0
python-dotenv==1.0.1
environs==11.0.0
django-admin-sortable2==2.2.1
requests==2.32.3
